refactor: extract share capability and download helpers to kit utils (OK-58192)
