+++ packages/kit/src/views/Perp/components/PositionShare/useShareActions.tsx
@@ -4,6 +4,7 @@ import { useIntl } from 'react-intl';
 import { Linking } from 'react-native';
 
 import { Button, Toast, useClipboard } from '@onekeyhq/components';
+import { shareImageOnDesktop } from '@onekeyhq/kit/src/utils/shareUtils';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 import MediaLibrary from '@onekeyhq/shared/src/modules3rdParty/expo-media-library';
 import Sharing from '@onekeyhq/shared/src/modules3rdParty/expo-sharing';
@@ -165,24 +166,10 @@ export function useShareActions(referralQrCodeUrl?: string) {
 
         await RNFS.unlink(filepath);
       } else if (platformEnv.isDesktop) {
-        // Electron has no navigator.share; on macOS the main process pops the
-        // native share picker (ShareMenu). Other desktop platforms have no
-        // system share, so fall back to saving the file.
-        const shared: boolean =
-          await globalThis.desktopApiProxy.system.shareImageFile({
-            base64Image,
-          });
-        if (!shared) {
-          const blob = await fetch(base64Image).then((r) => r.blob());
-          const url = URL.createObjectURL(blob);
-          const link = document.createElement('a');
-          link.href = url;
-          link.download = `onekey-position-${Date.now()}.png`;
-          document.body.appendChild(link);
-          link.click();
-          document.body.removeChild(link);
-          URL.revokeObjectURL(url);
-        }
+        await shareImageOnDesktop(
+          base64Image,
+          `onekey-position-${Date.now()}.png`,
+        );
       } else {
         // Web: Use Web Share API if available
         const byteString = atob(base64Image.split(',')[1]);
